working on creating routes for querying and need to rework query function to accept different queries
